# (WIP) PERF: improve .str perf for all-string values (about 2x-)