@jorisvandenbossche 2015-05-14
Where did the initial speed-up come from then?