@sinhrks 2015-05-17
The above difference exists on current master depending on dtypes handled. I've misunderstand it is caused be the logic i've changed, but not.